Added all the normal Guard functionality to the plugin.
I also added logic to try and find a Bundle & Guard based
project directory out of the open folders in the window
to run Guard in. If it can't find one then it throws an
error message saying it can't find one. This makes running
Guard generic enough that it would be valuable to other
people.