Added word_wrap and Guard auto termination.
I figured out how to enable native Sublime Text 2 view based word
wrapping in the output pane so I enabled that. It is a little
strange with respect to autoscrolling because the output is not
hard wrapped. I think that can be fixed by actually scrolling to
the size of the view like I used to rather than the first char in
each line. This will require further testing and will be resolved
in another commit. I also added a guard_wrapper command which is
now what is executed by the plugin. It simply wraps IO for the
guard command and watches for its STDIN to be closed. It does
this because when the wrappers STDIN closes that means that
Sublime Text 2 has exited either cleanly or by crash and we need
to stop Guard.