Added default key binding to show the guard view.
The default binding is super+shift+c. The reason that this command
of all of the guard commands is worthy of a key binding is because
the guard view hides when you do a find or when you simply press
escape. So, things that you do in the normal dev process may hide
the guard output and you need to be able to bring it back up
quickly. Hence, the key binding.