Start guard now set perms of internal scripts.
I added this because it seems that when the package gets installed
via the Package Control plugin it neglects to retain file permissions
as Git does by default when installed manually. Beyond that Git can
be configured in such a way to ignore the permissions. Therefore,
I decided to simply have the Python portion of the plugin set the perms
to 755 on both of the internal scripts every time start guard is run.
That way if they ever lose their perms it won't matter they will just
be correct such that they can be executed appropriately.