Is now relative to provided listener.
What does this mean. Well basically whenever an event is triggered
or a command is executed in Sublime Text 2 it provides the object
that received that thing (maybe the current window, etc.) and passes
that along. Now, this plugin is aware of this object and does things
relative to this object. This fixes #24.