Resolved syntax error identified in issue #27.
The big issue with issue #27 is that it wasn't identifying the end of
the running tests block. So, I added the keyword 'Done' to the ending
regex to make sure the next test is colored correctly. I also added
'backtrace' as an ending identifier to prevent backtraces at that level
from being colored. This resolves issue #27.