Fixed issue where begin dots running didn't match.
There were a couple cases I ran into in the wild since my dev team
changed our RSpec configuration a bit that caused the dots running
portion of the RSpec output to not match anymore. This fixes it so that
it is now matched and still matches the old pattern.